+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chess/dal/ChessFeedFilter.kt
@@ -20,62 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chess.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.quartz.nip64Chess.game.ChessGameEvent
-
-/**
- * Feed filter for NIP-64 Chess Game events (Kind 64)
- *
- * Filters the local cache to show only chess games,
- * respecting user's follow lists and hidden users.
- */
-class ChessFeedFilter(
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-chess-feed"
-
-    override fun limit() = 200
-
-    override fun feed(): List<Note> {
-        val params = buildFilterParams(account)
-
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it, params)
-            }
-
-        return sort(notes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val params = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) { acceptableEvent(it, params) }
-    }
-
-    fun acceptableEvent(
-        note: Note,
-        params: FilterByListParams,
-    ): Boolean {
-        val noteEvent = note.event
-
-        return (noteEvent is ChessGameEvent) &&
-            params.match(noteEvent, note.relays) &&
-            (params.isHiddenList || account.isAcceptable(note))
-    }
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            followLists = account.liveStoriesFollowLists.value,
-            hiddenUsers = account.hiddenUsers.flow.value,
-        )
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias ChessFeedFilter = com.vitorpamplona.amethyst.commons.ui.dal.ChessFeedFilter
