+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/polls/dal/ClosedPollsFeedFilter.kt
@@ -20,78 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.polls.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.TopFilter
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.amethyst.ui.dal.FilterByListParams
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
-import com.vitorpamplona.quartz.utils.TimeUtils
-
-class ClosedPollsFeedFilter(
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followList().code + "-closed"
-
-    override fun limit() = 200
-
-    fun followList(): TopFilter = account.settings.defaultPollsFollowList.value
-
-    fun TopFilter.isMuteList() = this is TopFilter.MuteList
-
-    fun TopFilter.isBlockList() = this is TopFilter.PeopleList && this.address == account.blockPeopleList.getBlockListAddress()
-
-    fun TopFilter.wantsToSeeNegativeStuff() = isMuteList() || isBlockList()
-
-    override fun showHiddenKey(): Boolean = followList().wantsToSeeNegativeStuff()
-
-    private fun isClosed(note: Note): Boolean {
-        val noteEvent = note.event
-        return when (noteEvent) {
-            is PollEvent -> {
-                noteEvent.hasEnded()
-            }
-
-            is ZapPollEvent -> {
-                val closedAt = noteEvent.closedAt()
-                closedAt != null && closedAt < TimeUtils.now()
-            }
-
-            else -> {
-                false
-            }
-        }
-    }
-
-    override fun feed(): List<Note> {
-        val params = buildFilterParams(account)
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                val noteEvent = it.event
-                (noteEvent is PollEvent || noteEvent is ZapPollEvent) && params.match(noteEvent, it.relays) && isClosed(it)
-            }
-        return sort(notes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    fun buildFilterParams(account: Account): FilterByListParams =
-        FilterByListParams.create(
-            account.livePollsFollowLists.value,
-            account.hiddenUsers.flow.value,
-        )
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        val params = buildFilterParams(account)
-
-        return collection.filterTo(HashSet()) {
-            val noteEvent = it.event
-            (noteEvent is PollEvent || noteEvent is ZapPollEvent) && params.match(noteEvent, it.relays) && isClosed(it)
-        }
-    }
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias ClosedPollsFeedFilter = com.vitorpamplona.amethyst.commons.ui.dal.ClosedPollsFeedFilter
